backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-anthropic
python-dotenv
python-multipart
